# public/swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Capstone - API - Documentation",
    "description": "API for user interface development",
    "version": "1.1",
    "contact": {
      "name": "mangox API Support",
      "email": "support@mangox.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/test-error/{code}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "code",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_test_error__code_",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/test-validation": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__api_v1_test_validation",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/test-auth": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_test_auth",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ]
      }
    },
    "/api/v1/test-user/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_test_user__id_",
        "tags": [
          "Common"
        ],
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://mangox-api.mangoads.com.vn",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:5555",
      "description": "Development Server"
    }
  ],
...
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}